Add new Page model
Page.cs contains fields for Id, Url, Title and Content. It also has a DateTimeCreated and a
DateTimeUpdated field, the first one indicates when the page is created, and it should not
be changed, the second one indicates the last moment changes were made to a page.
DateTimeDisplay indicates wether or not these dates should be displayed on the view. There's
also an Author field and an AuthorDisplay. The Author is currently a string, but it will be
changed to an Author object.