Change dataservices for di to scoped
Previously these dataservices were transient, but as written in the docs, services that use
Entity Framework, as the DataServices do, should use scoped lifetime.

"Entity Framework contexts should be added to the services container using the Scoped
lifetime. This is taken care of automatically if you use the helper methods as shown above.
Repositories that will make use of Entity Framework should use the same lifetime."