Apply big refactoring
- Move models and services to aCMS.Core
- Remove cashing from data services
- Rename aCMS to aCMS.Web
- Fix (almost) all the errors the previous three points created

Package restore in aCMS.Web currently not working, the reason for this is
currently unknown, but we have this (reaaally helpful (/s)) error message:
"Can not add property CommonMark.NET to Newtonsoft.Json.Linq.JObject.
Property with the same name already exists on object." If someone knows
what it means, please call me. I've tried removing and regenerating the
lockfile, but it didn't help...